targets/cryptography-secrets/ecb-mode-abuse/app/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pycryptodome==3.19.0
